requirements.txt
python-igraph
pycairo
numpy